Split multi-module test files to mirror their source paths
Codex flagged suites whose path credited one module while their tests
exercised others, leaving those modules uncredited by unit-tests-report:

- address-diff tests move out of address-lookup/coords-diff.test.ts into
  test/ui/client/admin/address-diff.test.ts (diffSpec shared via
  #test-utils/address-lookup-dom.ts)
- initLogisticsMapLoader tests move to logistics-map-loader.test.ts
- markdown-editor.test.ts splits into markdown-editor-setup.test.ts,
  markdown-editor-toolbar.test.ts, and markdown-editor-loader.test.ts,
  with the shared happy-dom harness in #test-utils/markdown-editor-dom.ts
- attendeeAdminPath tests move from db/attendees/kind.test.ts to
  test/shared/attendee-links.test.ts
- the /admin/settings/google-wallet cases move from wallet/google.test.ts
  to test/features/admin/settings-wallets.test.ts (wallet config fixtures
  shared via #test-utils/google-wallet.ts)
- adminListingRecalculatePage tests move from listings/form-pages.test.ts
  to listings/aggregates.test.ts
- public error-page tests move from reservations/og.test.ts to
  test/ui/templates/public/errors.test.ts
