+++ test/features/wallet/google.test.ts
@@ -1,32 +1,18 @@
 import { expect } from "@std/expect";
 import { beforeEach, it as test } from "@std/testing/bdd";
-import { handleRequest } from "#routes";
 import { settings } from "#shared/db/settings.ts";
-import {
-  expectFlashRedirect,
-  expectHtml,
-  expectRedirect,
-  testRequiresAuth,
-} from "#test-utils/assertions.ts";
+import { expectHtml, expectRedirect } from "#test-utils/assertions.ts";
 import { generateGoogleTestCreds } from "#test-utils/crypto.ts";
 import { describeWithEnv } from "#test-utils/db.ts";
 import {
   createTestAttendeeWithToken,
   fetchAliceTicketPageBody,
 } from "#test-utils/db-helpers/attendees.ts";
-import { setTestEnv } from "#test-utils/env.ts";
-import { awaitTestRequest, mockFormRequest } from "#test-utils/mocks.ts";
-import { loginAsAdmin } from "#test-utils/session.ts";
-
-/** Configure all Google Wallet settings in the database */
-const configureGoogleWallet = async () => {
-  const creds = generateGoogleTestCreds();
-  await Promise.all([
-    settings.update.googleWallet.issuerId(creds.issuerId),
-    settings.update.googleWallet.serviceAccountEmail(creds.serviceAccountEmail),
-    settings.update.googleWallet.serviceAccountKey(creds.serviceAccountKey),
-  ]);
-};
+import {
+  configureGoogleWallet,
+  setGoogleWalletEnvVars,
+} from "#test-utils/google-wallet.ts";
+import { awaitTestRequest } from "#test-utils/mocks.ts";
 
 describeWithEnv("google wallet route (/gwallet/:token)", { db: true }, () => {
   beforeEach(() => {
@@ -136,196 +122,6 @@ describeWithEnv("ticket view google wallet link", { db: true }, () => {
   });
 });
 
-describeWithEnv("POST /admin/settings/google-wallet", { db: true }, () => {
-  beforeEach(() => {
-    generateGoogleTestCreds();
-  });
-
-  testRequiresAuth("/admin/settings/google-wallet", {
-    body: {
-      google_wallet_issuer_id: "123",
-    },
-    method: "POST",
-  });
-
-  test("requires Issuer ID", async () => {
-    const { cookie, csrfToken } = await loginAsAdmin();
-
-    const response = await handleRequest(
-      mockFormRequest(
-        "/admin/settings/google-wallet",
-        {
-          csrf_token: csrfToken,
-          google_wallet_issuer_id: "",
-          google_wallet_service_account_email:
-            "test@test.iam.gserviceaccount.com",
-          google_wallet_service_account_key:
-            generateGoogleTestCreds().serviceAccountKey,
-        },
-        cookie,
-      ),
-    );
-    await expectFlashRedirect(
-      "/admin/settings-advanced?form=settings-google-wallet#settings-google-wallet",
-      expect.stringContaining("Issuer ID is required"),
-      false,
-    )(response);
-  });
-
-  test("requires Service Account Email", async () => {
-    const { cookie, csrfToken } = await loginAsAdmin();
-
-    const response = await handleRequest(
-      mockFormRequest(
-        "/admin/settings/google-wallet",
-        {
-          csrf_token: csrfToken,
-          google_wallet_issuer_id: "1234567890",
-          google_wallet_service_account_email: "",
-          google_wallet_service_account_key:
-            generateGoogleTestCreds().serviceAccountKey,
-        },
-        cookie,
-      ),
-    );
-    await expectFlashRedirect(
-      "/admin/settings-advanced?form=settings-google-wallet#settings-google-wallet",
-      expect.stringContaining("Service account email is required"),
-      false,
-    )(response);
-  });
-
-  test("requires private key on initial setup", async () => {
-    const { cookie, csrfToken } = await loginAsAdmin();
-
-    const response = await handleRequest(
-      mockFormRequest(
-        "/admin/settings/google-wallet",
-        {
-          csrf_token: csrfToken,
-          google_wallet_issuer_id: "1234567890",
-          google_wallet_service_account_email:
-            "test@test.iam.gserviceaccount.com",
-          google_wallet_service_account_key: "",
-        },
-        cookie,
-      ),
-    );
-    await expectFlashRedirect(
-      "/admin/settings-advanced?form=settings-google-wallet#settings-google-wallet",
-      expect.stringContaining("Service account private key is required"),
-      false,
-    )(response);
-  });
-
-  test("rejects invalid PEM private key", async () => {
-    const { cookie, csrfToken } = await loginAsAdmin();
-
-    const response = await handleRequest(
-      mockFormRequest(
-        "/admin/settings/google-wallet",
-        {
-          csrf_token: csrfToken,
-          google_wallet_issuer_id: "1234567890",
-          google_wallet_service_account_email:
-            "test@test.iam.gserviceaccount.com",
-          google_wallet_service_account_key: "not a valid key",
-        },
-        cookie,
-      ),
-    );
-    await expectFlashRedirect(
-      "/admin/settings-advanced?form=settings-google-wallet#settings-google-wallet",
-      expect.stringContaining(
-        "Service account private key is not a valid PEM private key",
-      ),
-      false,
-    )(response);
-  });
-
-  test("saves all settings successfully", async () => {
-    const { cookie, csrfToken } = await loginAsAdmin();
-
-    const response = await handleRequest(
-      mockFormRequest(
-        "/admin/settings/google-wallet",
-        {
-          csrf_token: csrfToken,
-          google_wallet_issuer_id: "1234567890",
-          google_wallet_service_account_email:
-            "test@test.iam.gserviceaccount.com",
-          google_wallet_service_account_key:
-            generateGoogleTestCreds().serviceAccountKey,
-        },
-        cookie,
-      ),
-    );
-
-    await expectFlashRedirect(
-      "/admin/settings-advanced?form=settings-google-wallet#settings-google-wallet",
-      "Google Wallet configuration updated",
-    )(response);
-
-    expect(settings.googleWallet.hasConfig).toBe(true);
-    expect(settings.googleWallet.issuerId).toBe("1234567890");
-    expect(settings.googleWallet.serviceAccountEmail).toBe(
-      "test@test.iam.gserviceaccount.com",
-    );
-  });
-
-  test("clears all settings when everything is empty", async () => {
-    await configureGoogleWallet();
-    expect(settings.googleWallet.hasConfig).toBe(true);
-
-    const { cookie, csrfToken } = await loginAsAdmin();
-    const response = await handleRequest(
-      mockFormRequest(
-        "/admin/settings/google-wallet",
-        {
-          csrf_token: csrfToken,
-          google_wallet_issuer_id: "",
-          google_wallet_service_account_email: "",
-          google_wallet_service_account_key: "",
-        },
-        cookie,
-      ),
-    );
-
-    await expectFlashRedirect(
-      "/admin/settings-advanced?form=settings-google-wallet#settings-google-wallet",
-      "Google Wallet configuration cleared",
-    )(response);
-    expect(settings.googleWallet.hasDbConfig).toBe(false);
-  });
-
-  test("shows Google Wallet section with values when configured", async () => {
-    await configureGoogleWallet();
-    const { cookie } = await loginAsAdmin();
-    const response = await awaitTestRequest("/admin/settings-advanced", {
-      cookie,
-    });
-    const body = await response.text();
-    expect(body).toContain("Google Wallet");
-    expect(body).toContain("google_wallet_issuer_id");
-    expect(body).toContain("1234567890");
-    expect(body).toContain("test@test-project.iam.gserviceaccount.com");
-    // Secret is masked
-    expect(body).toContain("••••••••");
-  });
-});
-
-/** Set all Google Wallet env vars and return restore function */
-const setGoogleWalletEnvVars = async () => {
-  generateGoogleTestCreds();
-  return setTestEnv({
-    GOOGLE_WALLET_ISSUER_ID: "9876543210",
-    GOOGLE_WALLET_SERVICE_ACCOUNT_EMAIL:
-      "env@env-project.iam.gserviceaccount.com",
-    GOOGLE_WALLET_SERVICE_ACCOUNT_KEY:
-      generateGoogleTestCreds().serviceAccountKey,
-  });
-};
-
 describeWithEnv(
   "googleWallet.getHostConfig",
   {
@@ -415,29 +211,6 @@ describeWithEnv(
       );
       const response = await awaitTestRequest(`/t/${token}`);
       await expectHtml(response, { contains: ["Google Wallet"] });
-    });
-
-    test("settings page shows host Google Wallet label when env vars configured", async () => {
-      await setGoogleWalletEnvVars();
-      const { cookie } = await loginAsAdmin();
-      const response = await awaitTestRequest("/admin/settings-advanced", {
-        cookie,
-      });
-      await expectHtml(response, {
-        contains: ["Host env (9876543210)", "Currently using"],
-      });
-    });
-
-    test("settings page shows overriding label when both DB and env configured", async () => {
-      await setGoogleWalletEnvVars();
-      await configureGoogleWallet();
-      const { cookie } = await loginAsAdmin();
-      const response = await awaitTestRequest("/admin/settings-advanced", {
-        cookie,
-      });
-      await expectHtml(response, {
-        contains: ["Host env (9876543210)", "Overriding"],
-      });
     });
   },
 );
