+++ test/ui/client/admin/logistics-map.test.ts
@@ -13,7 +13,6 @@ import {
   type MapLibrary,
   readMapPin,
 } from "#src/ui/client/admin/logistics-map.ts";
-import { initLogisticsMapLoader } from "#src/ui/client/admin/logistics-map-loader.ts";
 import { createDomInstaller } from "#test-utils/happy-dom.ts";
 
 const { installDom, cleanup } = createDomInstaller();
@@ -132,29 +131,5 @@ describe("initLogisticsMap", () => {
         .querySelector("[data-logistics-map]")!
         .hasAttribute("hidden"),
     ).toBe(true);
-  });
-});
-
-describe("initLogisticsMapLoader", () => {
-  test("injects the map bundle and stylesheet when the page has a map", () => {
-    const window = installDom(
-      '<script src="/admin.js?ts=99"></script><div data-logistics-map></div>',
-    );
-    initLogisticsMapLoader();
-    const head = window.document.head;
-    expect(head.querySelector("script")?.getAttribute("src")).toBe(
-      "/logistics-map.js?ts=99",
-    );
-    expect(head.querySelector("link")?.getAttribute("href")).toBe(
-      "/logistics-map.css?ts=99",
-    );
-    expect(head.querySelector("link")?.getAttribute("rel")).toBe("stylesheet");
-  });
-
-  test("loads nothing on pages without a map", () => {
-    const window = installDom("<div></div>");
-    initLogisticsMapLoader();
-    expect(window.document.head.querySelector("script")).toBeNull();
-    expect(window.document.head.querySelector("link")).toBeNull();
   });
 });
